Move button::interrupt into lib
