+++ src/lib.rs
@@ -1,4 +1,5 @@
 #![no_std]
 
 pub use stm32f3xx_hal::*;
+pub mod button;
 pub mod leds;
